7042 refactor common code to helper method. So, that it can be used commonly where ever needed AB#7042 AB#9122
