+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.API/Startup.cs
@@ -30,10 +30,6 @@ using System.Net.Http.Headers;
 using UKHO.ExchangeSetService.Common.Storage;
 using Microsoft.AspNetCore.Authorization;
 using UKHO.ExchangeSetService.Common.HealthCheck;
-using System.Net.Http;
-using Polly;
-using System.Net;
-using System.Threading.Tasks;
 using UKHO.ExchangeSetService.Common.Logging;
 
 namespace UKHO.ExchangeSetService.API
@@ -128,6 +124,8 @@ namespace UKHO.ExchangeSetService.API
 
             services.Configure<SalesCatalogueConfiguration>(configuration.GetSection("SalesCatalogue"));
 
+            var retryCount = Convert.ToInt32(configuration["RetryConfiguration:RetryCount"]);
+            var sleepDuration = Convert.ToDouble(configuration["RetryConfiguration:SleepDuration"]);
             services.AddHttpClient<ISalesCatalogueClient, SalesCatalogueClient>(client =>
                 {
                     client.BaseAddress = new Uri(configuration["SalesCatalogue:BaseUrl"]);
@@ -136,7 +134,7 @@ namespace UKHO.ExchangeSetService.API
                     client.DefaultRequestHeaders.UserAgent.Add(productHeaderValue);
                 }
             )
-            .AddHeaderPropagation().AddPolicyHandler((services, request) => GetRetryPolicy(services.GetService<ILogger<ISalesCatalogueClient>>(), "Sales Catalogue", EventIds.RetryHttpClientSCSRequest));
+            .AddHeaderPropagation().AddPolicyHandler((services, request) => CommonHelper.GetRetryPolicy(services.GetService<ILogger<ISalesCatalogueClient>>(), "Sales Catalogue", EventIds.RetryHttpClientSCSRequest, retryCount, sleepDuration));
 
             services.Configure<FileShareServiceConfiguration>(configuration.GetSection("FileShareService"));
             services.Configure<EssManagedIdentityConfiguration>(configuration.GetSection("ESSManagedIdentity"));
@@ -151,7 +149,7 @@ namespace UKHO.ExchangeSetService.API
                     client.DefaultRequestHeaders.UserAgent.Add(productHeaderValue);
                 }
             )
-            .AddHeaderPropagation().AddPolicyHandler((services, request) => GetRetryPolicy(services.GetService<ILogger<IFileShareServiceClient>>(), "File Share", EventIds.RetryHttpClientFSSRequest));
+            .AddHeaderPropagation().AddPolicyHandler((services, request) => CommonHelper.GetRetryPolicy(services.GetService<ILogger<IFileShareServiceClient>>(), "File Share", EventIds.RetryHttpClientFSSRequest, retryCount, sleepDuration));
             services.AddScoped<IFileSystemHelper, FileSystemHelper>();
             services.AddScoped<IFileShareService, FileShareService>();
             services.AddScoped<IProductDataService, ProductDataService>();
@@ -308,33 +306,6 @@ namespace UKHO.ExchangeSetService.API
 
             app.UseCorrelationIdMiddleware()
                .UseErrorLogging(loggerFactory);
-        }
-
-        private IAsyncPolicy<HttpResponseMessage> GetRetryPolicy(ILogger logger, string requestType, EventIds eventId)
-        {
-
-            var retryCount = Convert.ToInt32(configuration["RetryConfiguration:RetryCount"]);
-            var sleepDuration = Convert.ToDouble(configuration["RetryConfiguration:SleepDuration"]);
-            return Policy
-                .HandleResult<HttpResponseMessage>(r => r.StatusCode == HttpStatusCode.ServiceUnavailable || r.StatusCode == HttpStatusCode.TooManyRequests)
-                .OrResult(r => r.StatusCode == HttpStatusCode.ServiceUnavailable || r.StatusCode == HttpStatusCode.TooManyRequests)
-                .WaitAndRetryAsync(retryCount, (retryAttempt) =>
-                {
-                    return TimeSpan.FromSeconds(Math.Pow(sleepDuration, (retryAttempt - 1)));
-                }, async (response, timespan, retryAttempt, context) =>
-                {
-                    var retryAfterHeader = response.Result.Headers.FirstOrDefault(h => h.Key.ToLowerInvariant() == "retry-after");
-                    var correlationId = response.Result.RequestMessage.Headers.FirstOrDefault(h => h.Key.ToLowerInvariant() == "x-correlation-id");
-                    int retryAfter = 0;
-                    if (response.Result.StatusCode == HttpStatusCode.TooManyRequests && retryAfterHeader.Value != null && retryAfterHeader.Value.Any())
-                    {
-                        retryAfter = int.Parse(retryAfterHeader.Value.First());
-                        await Task.Delay(TimeSpan.FromMilliseconds(retryAfter));
-                    }
-                    logger
-                    .LogInformation(eventId.ToEventId(), "Re-trying {requestType} service with uri {RequestUri} request with delay {delay}ms and retry attempt {retry} with _X-Correlation-ID:{correlationId} as previous request was responded with {StatusCode}.",
-                    requestType, response.Result.RequestMessage.RequestUri, timespan.Add(TimeSpan.FromMilliseconds(retryAfter)).TotalMilliseconds, retryAttempt, correlationId.Value, response.Result.StatusCode);
-                });
         }
     }
 }
