+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.Common/Helpers/CommonHelper.cs
@@ -1,8 +1,15 @@
-﻿using System;
+﻿using Microsoft.Extensions.Logging;
+using Polly;
+using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
+using System.Linq;
+using System.Net;
+using System.Net.Http;
 using System.Security.Cryptography;
+using System.Threading.Tasks;
+using UKHO.ExchangeSetService.Common.Logging;
 using UKHO.ExchangeSetService.Common.Models.SalesCatalogue;
 
 namespace UKHO.ExchangeSetService.Common.Helpers
@@ -62,6 +69,30 @@ namespace UKHO.ExchangeSetService.Common.Helpers
                 }
             }
             return fileSize;
+        }
+
+        public static IAsyncPolicy<HttpResponseMessage> GetRetryPolicy(ILogger logger, string requestType, EventIds eventId, int retryCount, double sleepDuration)
+        {
+            return Policy
+                .HandleResult<HttpResponseMessage>(r => r.StatusCode == HttpStatusCode.ServiceUnavailable)
+                .OrResult(r => r.StatusCode == HttpStatusCode.TooManyRequests)
+                .WaitAndRetryAsync(retryCount, (retryAttempt) =>
+                {
+                    return TimeSpan.FromSeconds(Math.Pow(sleepDuration, (retryAttempt - 1)));
+                }, async (response, timespan, retryAttempt, context) =>
+                {
+                    var retryAfterHeader = response.Result.Headers.FirstOrDefault(h => h.Key.ToLowerInvariant() == "retry-after");
+                    var correlationId = response.Result.RequestMessage.Headers.FirstOrDefault(h => h.Key.ToLowerInvariant() == "x-correlation-id");
+                    int retryAfter = 0;
+                    if (response.Result.StatusCode == HttpStatusCode.TooManyRequests && retryAfterHeader.Value != null && retryAfterHeader.Value.Any())
+                    {
+                        retryAfter = int.Parse(retryAfterHeader.Value.First());
+                        await Task.Delay(TimeSpan.FromMilliseconds(retryAfter));
+                    }
+                    logger
+                    .LogInformation(eventId.ToEventId(), "Re-trying {requestType} service with uri {RequestUri} request with delay {delay}ms and retry attempt {retry} with _X-Correlation-ID:{correlationId} as previous request was responded with {StatusCode}.",
+                    requestType, response.Result.RequestMessage.RequestUri, timespan.Add(TimeSpan.FromMilliseconds(retryAfter)).TotalMilliseconds, retryAttempt, correlationId.Value, response.Result.StatusCode);
+                });
         }
     }
 }
