+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.FulfilmentService/UKHO.ExchangeSetService.FulfilmentService.csproj
@@ -34,7 +34,6 @@
 		<PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.9" />
 		<PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="5.0.0" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="3.1.17" />
-    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="3.1.17" />
 		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
 		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="3.1.9" />
 		<PackageReference Include="runtime.native.System.Net.Security" Version="4.3.1" />
